Sat, Jan 21, 2023, 11:51 AM +00:00

+++ credit/credit.c
@@ -46,7 +46,7 @@ int get_creditnum(void)
 
     int j;
 
-    int i;
+    float i;
 
     int valid;
